+++ src/arch/armv8m.rs
@@ -1,0 +1,150 @@
+// Generated asm:
+// - armv8-m baseline https://godbolt.org/z/oYKWv7aYT
+// - armv8-m mainline https://godbolt.org/z/an9TYYjvn
+
+use core::{arch::asm, mem::MaybeUninit, sync::atomic::Ordering};
+
+use crate::raw::{AtomicLoad, AtomicStore, AtomicSwap};
+
+macro_rules! atomic {
+    ($int_type:ident, $asm_suffix:tt) => {
+        impl AtomicLoad for $int_type {
+            #[inline]
+            unsafe fn atomic_load(
+                src: *const MaybeUninit<Self>,
+                out: *mut MaybeUninit<Self>,
+                order: Ordering,
+            ) {
+                // clippy bug that does not recognize safety comments inside macros.
+                #[allow(clippy::undocumented_unsafe_blocks)]
+                // SAFETY: the caller must uphold the safety contract for `atomic_load`.
+                unsafe {
+                    macro_rules! atomic_load {
+                        ($acq:tt) => {
+                            asm!(
+                                // (atomic) load from src to tmp
+                                concat!("ld", $acq, $asm_suffix, " {tmp}, [{src}]"),
+                                // store tmp to out
+                                concat!("str", $asm_suffix, " {tmp}, [{out}]"),
+                                src = in(reg) src,
+                                out = inout(reg) out => _,
+                                tmp = lateout(reg) _,
+                                options(nostack),
+                            )
+                        };
+                    }
+                    match order {
+                        Ordering::Relaxed => atomic_load!("r"),
+                        // Acquire and SeqCst loads are equivalent.
+                        Ordering::Acquire | Ordering::SeqCst => atomic_load!("a"),
+                        _ => crate::utils::release_unreachable_unchecked(),
+                    }
+                }
+            }
+        }
+        impl AtomicStore for $int_type {
+            #[inline]
+            unsafe fn atomic_store(
+                dst: *mut MaybeUninit<Self>,
+                val: *const MaybeUninit<Self>,
+                order: Ordering,
+            ) {
+                // clippy bug that does not recognize safety comments inside macros.
+                #[allow(clippy::undocumented_unsafe_blocks)]
+                // SAFETY: the caller must uphold the safety contract for `atomic_store`.
+                unsafe {
+                    macro_rules! atomic_store {
+                        ($rel:tt) => {
+                            asm!(
+                                // load from val to tmp
+                                concat!("ldr", $asm_suffix, " {tmp}, [{val}]"),
+                                // (atomic) store tmp to dst
+                                concat!("st", $rel, $asm_suffix, " {tmp}, [{dst}]"),
+                                dst = inout(reg) dst => _,
+                                val = in(reg) val,
+                                tmp = lateout(reg) _,
+                                options(nostack),
+                            )
+                        };
+                    }
+                    match order {
+                        Ordering::Relaxed => atomic_store!("r"),
+                        // Release and SeqCst stores are equivalent.
+                        Ordering::Release | Ordering::SeqCst => atomic_store!("l"),
+                        _ => crate::utils::release_unreachable_unchecked(),
+                    }
+                }
+            }
+        }
+        impl AtomicSwap for $int_type {
+            #[inline]
+            unsafe fn atomic_swap(
+                dst: *mut MaybeUninit<Self>,
+                val: *const MaybeUninit<Self>,
+                out: *mut MaybeUninit<Self>,
+                order: Ordering,
+            ) {
+                // clippy bug that does not recognize safety comments inside macros.
+                #[allow(clippy::undocumented_unsafe_blocks)]
+                // SAFETY: the caller must uphold the safety contract for `atomic_swap`.
+                unsafe {
+                    macro_rules! atomic_swap {
+                        ($acq:tt, $rel:tt) => {
+                            asm!(
+                                // load from val to val_tmp
+                                concat!("ldr", $asm_suffix, " {val_tmp}, [{val}]"),
+                                // (atomic) swap
+                                "2:",
+                                    // load from dst to out_tmp
+                                    concat!("ld", $acq, "ex", $asm_suffix, " {out_tmp}, [{dst}]"),
+                                    // store val to dst
+                                    concat!("st", $rel, "ex", $asm_suffix, " {r}, {val_tmp}, [{dst}]"),
+                                    // 0 if the store was successful, 1 if no store was performed
+                                    "cmp {r}, 0x0",
+                                    "bne 2b",
+                                // store out_tmp to out
+                                concat!("str", $asm_suffix, " {out_tmp}, [{out}]"),
+                                dst = inout(reg) dst => _,
+                                val = in(reg) val,
+                                out = inout(reg) out => _,
+                                r = lateout(reg) _,
+                                out_tmp = lateout(reg) _,
+                                val_tmp = lateout(reg) _,
+                                options(nostack),
+                            )
+                        };
+                    }
+                    match order {
+                        Ordering::Relaxed => atomic_swap!("r", "r"),
+                        Ordering::Acquire => atomic_swap!("a", "r"),
+                        Ordering::Release => atomic_swap!("r", "l"),
+                        // AcqRel and SeqCst swaps are equivalent.
+                        Ordering::AcqRel | Ordering::SeqCst => atomic_swap!("a", "l"),
+                        _ => crate::utils::release_unreachable_unchecked(),
+                    }
+                }
+            }
+        }
+    };
+}
+
+atomic!(i8, "b");
+atomic!(u8, "b");
+atomic!(i16, "h");
+atomic!(u16, "h");
+atomic!(i32, "");
+atomic!(u32, "");
+atomic!(isize, "");
+atomic!(usize, "");
+
+#[cfg(test)]
+mod tests {
+    test_atomic!(isize);
+    test_atomic!(usize);
+    test_atomic!(i8);
+    test_atomic!(u8);
+    test_atomic!(i16);
+    test_atomic!(u16);
+    test_atomic!(i32);
+    test_atomic!(u32);
+}