+++ src/arch/mod.rs
@@ -8,16 +8,7 @@
 #[cfg(not(any(
     target_arch = "x86",
     target_arch = "x86_64",
-    all(
-        target_arch = "arm",
-        any(target_feature = "v7", atomic_maybe_uninit_target_feature_v7),
-        any(
-            target_feature = "aclass",
-            atomic_maybe_uninit_target_feature_aclass,
-            target_feature = "rclass",
-            atomic_maybe_uninit_target_feature_rclass,
-        )
-    ),
+    all(target_arch = "arm", any(target_feature = "v7", atomic_maybe_uninit_target_feature_v7)),
     target_arch = "aarch64",
     target_arch = "riscv32",
     target_arch = "riscv64"
@@ -26,17 +17,12 @@ compile_error!("this target is not supported yet");
 
 #[cfg(target_arch = "aarch64")]
 mod aarch64;
-#[cfg(all(
-    target_arch = "arm",
-    any(target_feature = "v7", atomic_maybe_uninit_target_feature_v7),
-    any(
-        target_feature = "aclass",
-        atomic_maybe_uninit_target_feature_aclass,
-        target_feature = "rclass",
-        atomic_maybe_uninit_target_feature_rclass,
-    )
-))]
-mod arm;
+#[cfg(all(target_arch = "arm", any(target_feature = "v7", atomic_maybe_uninit_target_feature_v7)))]
+#[cfg(not(any(target_feature = "v8", atomic_maybe_uninit_target_feature_v8)))]
+mod armv7;
+#[cfg(all(target_arch = "arm", any(target_feature = "v8", atomic_maybe_uninit_target_feature_v8)))]
+#[cfg(any(target_feature = "mclass", atomic_maybe_uninit_target_feature_mclass))]
+mod armv8m;
 #[cfg(any(target_arch = "riscv32", target_arch = "riscv64"))]
 mod riscv;
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]